+++ packages/multichain-account-service/src/MultichainAccountGroup.test.ts
@@ -5,6 +5,7 @@ import {
   toMultichainAccountGroupId,
   toMultichainAccountWalletId,
 } from '@metamask/account-api';
+import type { Messenger } from '@metamask/base-controller';
 import { EthScope, SolScope } from '@metamask/keyring-api';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
 
@@ -18,13 +19,20 @@ import {
   MOCK_WALLET_1_ENTROPY_SOURCE,
   MOCK_WALLET_1_EVM_ACCOUNT,
   MOCK_WALLET_1_SOL_ACCOUNT,
-  setupAccountProvider,
+  setupNamedAccountProvider,
   getMultichainAccountServiceMessenger,
   getRootMessenger,
 } from './tests';
+import type {
+  AllowedActions,
+  AllowedEvents,
+  MultichainAccountServiceActions,
+  MultichainAccountServiceEvents,
+} from './types';
 
 function setup({
   groupIndex = 0,
+  messenger = getRootMessenger(),
   accounts = [
     [MOCK_WALLET_1_EVM_ACCOUNT],
     [
@@ -34,26 +42,33 @@ function setup({
       MOCK_SNAP_ACCOUNT_2, // Non-BIP-44 account.
     ],
   ],
-}: { groupIndex?: number; accounts?: InternalAccount[][] } = {}): {
+}: {
+  groupIndex?: number;
+  messenger?: Messenger<
+    MultichainAccountServiceActions | AllowedActions,
+    MultichainAccountServiceEvents | AllowedEvents
+  >;
+  accounts?: InternalAccount[][];
+} = {}): {
   wallet: MultichainAccountWallet<Bip44Account<InternalAccount>>;
   group: MultichainAccountGroup<Bip44Account<InternalAccount>>;
   providers: MockAccountProvider[];
 } {
   const providers = accounts.map((providerAccounts) => {
-    return setupAccountProvider({ accounts: providerAccounts });
+    return setupNamedAccountProvider({ accounts: providerAccounts });
   });
 
   const wallet = new MultichainAccountWallet<Bip44Account<InternalAccount>>({
-    providers,
     entropySource: MOCK_WALLET_1_ENTROPY_SOURCE,
-    messenger: getMultichainAccountServiceMessenger(getRootMessenger()),
+    messenger: getMultichainAccountServiceMessenger(messenger),
+    providers,
   });
 
   const group = new MultichainAccountGroup({
     wallet,
     groupIndex,
     providers,
-    messenger: getMultichainAccountServiceMessenger(getRootMessenger()),
+    messenger: getMultichainAccountServiceMessenger(messenger),
   });
 
   return { wallet, group, providers };